# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: hyve API
  description: |
    hyve's HTTP API + auth layer — see HYVE-CONTROLLER-ARCHITECTURE-PLAN.md's
    Phase 6. A thin, authorized front door onto the ClusterDefinition/
    Template/Workflow/Module CRDs the controller already reconciles —
    cluster mode never requires this API, plain kubectl against the CRDs
    always works. Organization/environment/RBAC-binding/session data lives
    in a separate Postgres/SQLite store (internal/orgdb), not Kubernetes —
    see HYVE-ORGANIZATION-MODEL-IMPLEMENTATION-PLAN.md, nexus-config/docs.

    Everything under /api/* requires a session: log in via /auth/login,
    then send the returned accessToken as `Authorization: Bearer
    <accessToken>` on every subsequent request. accessToken is short-lived
    (see AccessTokenTTL) — exchange sessionToken for a new one via
    /auth/refresh as needed, without re-authenticating with a
    username/password. A session is revocable server-side (backed by a
    Session row) via /auth/logout — unlike accessToken, which is a
    stateless signed value that can't be invalidated early.
  version: "0.1.0"
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: Authenticate with a local (username/password) identity and receive a new session
      description: |
        Creates a Session row (the revocable record /auth/logout
        deletes and /auth/refresh checks) and returns both halves of the
        resulting credential — see LoginResponse.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Missing username/password or malformed body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid username or password (the same error either way — login never reveals which usernames exist)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/refresh:
    post:
      tags: [auth]
      summary: Exchange a still-valid session token for a new access token
      description: |
        Does not require the Authorization header — that's the whole
        point, it must still work once the access token it's replacing has
        already expired. Looks up the Session row named by sessionToken's
        id half and checks the presented secret half against its stored
        hash. The session token itself is never reissued by this call —
        only a fresh access token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionTokenRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "400":
          description: Malformed session token or request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Session not found, expired, or the presented secret didn't match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/logout:
    post:
      tags: [auth]
      summary: Revoke a session
      security: []
      description: |
        Deletes the Session row named by the presented session token — real,
        immediate revocation: the very next /auth/refresh against it fails.
        Any access token already cached from it keeps working until its own
        short TTL lapses regardless (see AccessTokenTTL) — there's no
        cheaper way to invalidate an already-issued one. Best-effort and
        always returns 200: a missing/malformed session token or an
        already-gone session isn't an error from the caller's perspective,
        since the end state is identical either way.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionTokenRequest"
      responses:
        "200":
          description: OK
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: accessToken returned by POST /auth/login or /auth/refresh
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    LoginRequest:
      type: object
      properties:
        username: { type: string }
        password: { type: string, format: password }
      required: [username, password]
    LoginResponse:
      type: object
      description: |
        Two distinct credentials, not one — see orgdb.Session's own design
        rationale for why. accessToken is what's sent on every /api/*
        request (stateless, short-lived, verified locally by the server).
        sessionToken is the longer-lived, revocable credential — shaped
        "<id>.<secret>" — POST /auth/refresh consumes to mint a new
        accessToken without ever re-sending a username/password. Store
        both; a client that discards sessionToken can't silently renew and
        will need a fresh login once accessToken expires.
      properties:
        accessToken: { type: string }
        accessTokenExpiresAt: { type: string, format: date-time }
        sessionToken: { type: string }
        sessionExpiresAt: { type: string, format: date-time }
      required: [accessToken, accessTokenExpiresAt, sessionToken, sessionExpiresAt]
    SessionTokenRequest:
      type: object
      description: Shared request body for POST /auth/refresh and POST /auth/logout.
      properties:
        sessionToken: { type: string, description: "The \"<id>.<secret>\" value from LoginResponse.sessionToken" }
      required: [sessionToken]
    RefreshResponse:
      type: object
      properties:
        accessToken: { type: string }
        accessTokenExpiresAt: { type: string, format: date-time }
      required: [accessToken, accessTokenExpiresAt]
...